Create BacktestingSyntheticData
While backtesting on historical data is a key part of developing a trading strategy, it also has some limitations that are important to be aware of. Synthetic data is data that is artificially generated. It can be generated so as to try to mimic certain properties of a real historical dataset, or it can be generated to test your strategy against hypothetical scenarios and asset behaviour not seen in the historical data.